Gets local administrative units from Eurostat's website

Source: https://ec.europa.eu/eurostat/web/gisco/geodata/reference-data/administrative-units-statistical-units/lau#lau18

ll_get_lau_eu(
  gisco_id = NULL,
  name = NULL,
  year = 2020,
  silent = FALSE,
  lau_sf = NULL,
  fallback = TRUE
)

Arguments

gisco_id

Gisco identifier of the relevant administrative unit. If given, takes precedence over name.

name

Name of the local administrative unit in the local language. Use gisco_id whenever possible, as names of local administrative units are not unique, e.g. there are 11 "Neuenkirchen" in the dataset. If both name and gisco_id are NULL, then it returns all municipalities.

year

Year of mapping, defaults to most recent (2020). Available starting with 2011.

silent

Defaults to FALSE. If TRUE, hides copyright notice. Useful e.g. when using this in reports or in loops. The copyright notice must still be shown where the final output is used.

lau_sf

sf object, exactly such as the one that would be returned by ll_get_lau_eu(). Used to speed-up computation when bulk processing.

Value

European LAU in sf format

Examples

ll_set_folder("~/R/")
#> [1] "~/R/"
ll_get_lau_eu()
#> ℹ © EuroGeographics for the administrative boundaries
#> Error in if (fallback == TRUE & (gisco_id %in% ll_codes$id)) {    code_row_df <- ll_codes %>% dplyr::filter(gisco_id == .data$id)    if (code_row_df$source == "ll_get_lau_eu()") {    }    else if (code_row_df$source == "ll_get_nuts_eu(level = 3)") {        usethis::ui_warn("Falling back on `ll_get_nuts_eu()`. Refer to original function for more options.")        return(ll_get_nuts_eu(nuts_id = gisco_id, level = 3))    }    else if (code_row_df$source == "ll_get_gadm(geo = 'UKR', level = 1)") {        lau_sf <- ll_get_gadm(geo = "UKR", level = 1) %>% dplyr::mutate(GISCO_ID = stringr::str_c("UA_",             GID_1))    }    else if (code_row_df$source == "ll_get_gadm(geo = 'UKR', level = 2)") {        lau_sf <- ll_get_gadm(geo = "UKR", level = 2) %>% dplyr::mutate(GISCO_ID = stringr::str_c("UA_",             GID_2))    }    else if (code_row_df$source == "ll_get_gadm(geo = 'BIH', level = 2)") {        lau_sf <- ll_get_gadm(geo = "BIH", level = 2) %>% dplyr::mutate(GISCO_ID = stringr::str_c("BA_",             GID_2))    }    else if (code_row_df$source == "ll_get_gadm(geo = 'BIH', level = 3)") {        lau_sf <- ll_get_gadm(geo = "BIH", level = 3) %>% dplyr::mutate(GISCO_ID = stringr::str_c("BA_",             GID_3))    }    else if (code_row_df$source == "ll_get_adm_ocha(geo = 'MD', level = 1)") {        lau_sf <- ll_get_adm_ocha(geo = "MD", level = 1) %>%             dplyr::mutate(GISCO_ID = ADM1_PCODE)    }    else if (code_row_df$source == "ll_get_gadm(geo = 'RS', level = 1)") {        lau_sf <- ll_get_gadm(geo = "RS", level = 1) %>% dplyr::mutate(GISCO_ID = stringr::str_c("RS_",             GID_3))    }    else if (code_row_df$source == "ll_get_gadm(geo = 'XKO', level = 1)") {        lau_sf <- ll_get_gadm(geo = "XKO", level = 1) %>% dplyr::mutate(GISCO_ID = stringr::str_c("XK_",             GID_1))    }    else if (code_row_df$source == "ll_get_gadm(geo = 'XKO', level = 2)") {        lau_sf <- ll_get_gadm(geo = "XKO", level = 2) %>% dplyr::mutate(GISCO_ID = stringr::str_c("XK_",             GID_2))    }}: argument is of length zero