Gets local administrative units from Eurostat's website

Source: https://ec.europa.eu/eurostat/web/gisco/geodata/reference-data/administrative-units-statistical-units/lau#lau18

Usage

ll_get_lau_eu(
  gisco_id = NULL,
  name = NULL,
  year = 2020,
  silent = FALSE,
  lau_sf = NULL,
  fallback = TRUE
)

Arguments

gisco_id

Gisco identifier of the relevant administrative unit. If given, takes precedence over name.

name

Name of the local administrative unit in the local language. Use gisco_id whenever possible, as names of local administrative units are not unique, e.g. there are 11 "Neuenkirchen" in the dataset. If both name and gisco_id are NULL, then it returns all municipalities.

year

Year of mapping, defaults to most recent (2020). Available starting with 2011.

silent

Defaults to FALSE. If TRUE, hides copyright notice. Useful e.g. when using this in reports or in loops. The copyright notice must still be shown where the final output is used.

lau_sf

sf object, exactly such as the one that would be returned by ll_get_lau_eu(). Used to speed-up computation when bulk processing.

Value

European LAU in sf format

Examples

ll_set_folder("~/R/")
#> [1] "~/R/"
ll_get_lau_eu()
#> ℹ © EuroGeographics for the administrative boundaries
#> Simple feature collection with 98613 features and 9 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -63.15372 ymin: -21.38881 xmax: 55.83573 ymax: 71.1853
#> Geodetic CRS:  WGS 84
#> # A tibble: 98,613 × 10
#>    GISCO_ID  CNTR_CODE LAU_ID LAU_NAME       POP_2…¹ POP_D…² AREA_…³  YEAR FID  
#>  * <chr>     <chr>     <chr>  <chr>            <int>   <dbl>   <dbl> <int> <chr>
#>  1 CZ_579475 CZ        579475 Libňatov           363   62.2     5.83  2020 CZ_5…
#>  2 CZ_579483 CZ        579483 Libotov            178   41.8     4.26  2020 CZ_5…
#>  3 CZ_579491 CZ        579491 Vranov             181   37.4     4.84  2020 CZ_5…
#>  4 CZ_579505 CZ        579505 Malá Úpa           142    5.32   26.7   2020 CZ_5…
#>  5 AT_41004  AT        41004  Eggendorf im …    1041  116.      8.94  2020 AT_4…
#>  6 CZ_579513 CZ        579513 Malé Svatoňov…    1509  222.      6.81  2020 CZ_5…
#>  7 CZ_579530 CZ        579530 Maršov u Úpice     191   57.9     3.30  2020 CZ_5…
#>  8 CZ_579548 CZ        579548 Mladé Buky        2308   86.4    26.7   2020 CZ_5…
#>  9 CZ_579556 CZ        579556 Mostek            1191   88.8    13.4   2020 CZ_5…
#> 10 CZ_579564 CZ        579564 Nemojov            761   93.7     8.12  2020 CZ_5…
#> # … with 98,603 more rows, 1 more variable: geometry <MULTIPOLYGON [°]>, and
#> #   abbreviated variable names ¹​POP_2020, ²​POP_DENS_2, ³​AREA_KM2